Added test for filterResources
